.deploy_builds/36eba25bbd/requirements.txt
# Auto-generated for framework=tensorflow, target CPython 3.11
# Pin versions in production as needed.
fastapi>=0.110.0
joblib>=1.3.0
numpy==1.26.4
pydantic>=2.0.0
tensorflow-cpu==2.15.0
uvicorn[standard]>=0.27.0
